Alterations to keep MPX resistance below 200 ohms
The MPX only allows a resistance of < 250 ohms. I've made the following alterations to keep the LTV-846 resistance to MPX ~200 ohms. The resistors between the LTV-846 and the MPX were there only to reduce current through the LTV. In conditions I've observed, the LTV-846 should be under it's maximum ratings without the resistors. If the current draw through the LTV-846 spikes for some reason, there is no risk to the MPX or STM32.
Removing these resistors should have no detrimental impact.